proto: rename module to record_type_set
